enhancement 646 - Added UI Theme setting, added @Mandar css as first option.  Looking at other ways to make this more accessible. Moved settings into a settings-helpers.js file to clean up swffg-main.js file.
