+++ modules/settings/settings-helpers.js
@@ -1,0 +1,210 @@
+import DataImporter from "../importer/data-importer.js";
+import SWAImporter from "../importer/swa-importer.js";
+
+export default class SettingsHelpers {
+  // Initialize System Settings after the Init Hook
+  static initLevelSettings() {
+    // System Migration Version
+    game.settings.register("starwarsffg", "systemMigrationVersion", {
+      name: "Current Version",
+      scope: "world",
+      default: null,
+      config: false,
+      type: String,
+    });
+
+    // Register dice theme setting
+    game.settings.register("starwarsffg", "dicetheme", {
+      name: game.i18n.localize("SWFFG.SettingsDiceTheme"),
+      hint: game.i18n.localize("SWFFG.SettingsDiceThemeHint"),
+      scope: "world",
+      config: true,
+      default: "starwars",
+      type: String,
+      onChange: (rule) => window.location.reload(),
+      choices: {
+        starwars: "starwars",
+        genesys: "genesys",
+      },
+    });
+
+    // Enable auto Soak calculation
+    game.settings.register("starwarsffg", "enableSoakCalc", {
+      name: game.i18n.localize("SWFFG.EnableSoakCalc"),
+      hint: game.i18n.localize("SWFFG.EnableSoakCalcHint"),
+      scope: "world",
+      config: true,
+      default: true,
+      type: Boolean,
+      onChange: (rule) => window.location.reload(),
+    });
+
+    // Register grouping talents so people can let them be ordered by purchase history
+    game.settings.register("starwarsffg", "talentSorting", {
+      name: game.i18n.localize("SWFFG.EnableSortTalentsByActivationGlobal"),
+      hint: game.i18n.localize("SWFFG.EnableSortTalentsByActivationHint"),
+      scope: "world",
+      config: true,
+      default: false,
+      type: Boolean,
+      onChange: (rule) => window.location.reload(),
+    });
+
+    // Register skill sorting by localised value setting
+    game.settings.register("starwarsffg", "skillSorting", {
+      name: game.i18n.localize("SWFFG.SettingsSkillSorting"),
+      hint: game.i18n.localize("SWFFG.SettingsSkillSortingHint"),
+      scope: "world",
+      config: true,
+      default: false,
+      type: Boolean,
+      onChange: (rule) => window.location.reload(),
+    });
+
+    // Register setting for group manager Player Character List display mode
+    game.settings.register("starwarsffg", "pcListMode", {
+      name: game.i18n.localize("SWFFG.SettingsPCListMode"),
+      hint: game.i18n.localize("SWFFG.SettingsPCListModeHint"),
+      scope: "world",
+      config: true,
+      default: "active",
+      type: String,
+      choices: {
+        active: game.i18n.localize("SWFFG.SettingsPCListModeActive"),
+        owned: game.i18n.localize("SWFFG.SettingsPCListModeOwned"),
+      },
+      onChange: (rule) => {
+        const groupmanager = canvas?.groupmanager?.window;
+        if (groupmanager) {
+          groupmanager.render();
+        }
+      },
+    });
+
+    // Register placeholder settings to store Destiny Pool values for the group manager.
+    game.settings.register("starwarsffg", "dPoolLight", {
+      name: "Destiny Pool Light",
+      scope: "world",
+      default: 0,
+      config: false,
+      type: Number,
+      onChange: (rule) => {
+        const groupmanager = canvas?.groupmanager?.window;
+        if (groupmanager) {
+          groupmanager.render();
+        }
+        let destinyLight = game.settings.get("starwarsffg", "dPoolLight");
+        document.getElementById("destinyLight").setAttribute("data-value", destinyLight);
+        document.getElementById("destinyLight").innerHTML = destinyLight + `<span>${game.i18n.localize("SWFFG.Lightside")}</span>`;
+      },
+    });
+    game.settings.register("starwarsffg", "dPoolDark", {
+      name: "Destiny Pool Dark",
+      scope: "world",
+      default: 0,
+      config: false,
+      type: Number,
+      onChange: (rule) => {
+        const groupmanager = canvas?.groupmanager?.window;
+        if (groupmanager) {
+          groupmanager.render();
+        }
+        let destinyDark = game.settings.get("starwarsffg", "dPoolDark");
+        document.getElementById("destinyDark").setAttribute("data-value", destinyDark);
+        document.getElementById("destinyDark").innerHTML = destinyDark + `<span>${game.i18n.localize("SWFFG.Darkside")}</span>`;
+      },
+    });
+
+    // OggDude Importer Control Menu
+    game.settings.registerMenu("starwarsffg", "odImporter", {
+      name: game.i18n.localize("SWFFG.SettingsOggDudeImporter"),
+      hint: game.i18n.localize("SWFFG.SettingsOggDudeImporterHint"),
+      label: game.i18n.localize("SWFFG.SettingsOggDudeImporterLabel"),
+      icon: "fas fa-file-import",
+      type: DataImporter,
+      restricted: true,
+    });
+    game.settings.register("starwarsffg", "odImporter", {
+      name: "Item Importer",
+      scope: "world",
+      default: {},
+      config: false,
+      default: {},
+      type: Object,
+    });
+
+    // SWA Importer Control Menu
+    game.settings.registerMenu("starwarsffg", "swaImporter", {
+      name: game.i18n.localize("SWFFG.SettingsSWAdversariesImporter"),
+      hint: game.i18n.localize("SWFFG.SettingsSWAdversariesImporterHint"),
+      label: game.i18n.localize("SWFFG.SettingsSWAdversariesImporterLabel"),
+      icon: "fas fa-file-import",
+      type: SWAImporter,
+      restricted: true,
+    });
+    game.settings.register("starwarsffg", "swaImporter", {
+      name: "Adversaries Importer",
+      scope: "world",
+      default: {},
+      config: false,
+      default: {},
+      type: Object,
+    });
+
+    // Enable debug messages in console
+    game.settings.register("starwarsffg", "enableDebug", {
+      name: game.i18n.localize("SWFFG.EnableDebug"),
+      hint: game.i18n.localize("SWFFG.EnableDebugHint"),
+      scope: "world",
+      config: true,
+      default: false,
+      type: Boolean,
+      onChange: (rule) => window.location.reload(),
+    });
+
+    game.settings.register("starwarsffg", "uitheme", {
+      name: game.i18n.localize("SWFFG.SettingsUITheme"),
+      hint: game.i18n.localize("SWFFG.SettingsUIsThemeHint"),
+      scope: "world",
+      config: true,
+      default: "default",
+      type: String,
+      onChange: (rule) => window.location.reload(),
+      choices: {
+        default: "Default",
+        mandar: "Mandar",
+      },
+    });
+  }
+
+  // Initialize System Settings after the Ready Hook
+  static readyLevelSetting() {
+    // Allow Users to Roll Audio
+    game.settings.register("starwarsffg", "allowUsersAddRollAudio", {
+      name: game.i18n.localize("SWFFG.EnableRollAudio"),
+      hint: game.i18n.localize("SWFFG.EnableRollAudioHint"),
+      scope: "world",
+      default: false,
+      config: true,
+      type: Boolean,
+    });
+
+    // generate a list of playlists
+    const playlists = {};
+    playlists["None"] = "";
+    game.playlists.entries.forEach((playlist, index) => {
+      playlists[playlist.id] = `${index}-${playlist.data.name}`;
+    });
+
+    // Playlist users can user for audio
+    game.settings.register("starwarsffg", "allowUsersAddRollAudioPlaylist", {
+      name: game.i18n.localize("SWFFG.EnableRollAudioPlaylist"),
+      hint: game.i18n.localize("SWFFG.EnableRollAudioPlaylistHint"),
+      scope: "world",
+      default: "None",
+      config: true,
+      type: String,
+      choices: playlists,
+    });
+  }
+}
