+++ modules/swffg-main.js
@@ -19,8 +19,7 @@ import { DicePoolFFG, RollFFG } from "./dice-pool-ffg.js";
 import { GroupManagerLayer } from "./groupmanager-ffg.js";
 import { GroupManager } from "./groupmanager-ffg.js";
 import PopoutEditor from "./popout-editor.js";
-import DataImporter from "./importer/data-importer.js";
-import SWAImporter from "./importer/swa-importer.js";
+
 import CharacterImporter from "./importer/character-importer.js";
 import DiceHelpers from "./helpers/dice-helpers.js";
 import Helpers from "./helpers/common.js";
@@ -28,6 +27,7 @@ import TemplateHelpers from "./helpers/partial-templates.js";
 import SkillListImporter from "./importer/skills-list-importer.js";
 import DestinyTracker from "./ffg-destiny-tracker.js";
 import { defaultSkillArrayString } from "./config/ffg-skillslist.js";
+import SettingsHelpers from "./settings/settings-helpers.js";
 
 // Import Dice Types
 import { AbilityDie, BoostDie, ChallengeDie, DifficultyDie, ForceDie, ProficiencyDie, SetbackDie } from "./dice-pool-ffg.js";
@@ -115,13 +115,20 @@ Hooks.once("init", async function () {
   // Load character templates so that dynamic skills lists work correctly
   loadTemplates(["systems/starwarsffg/templates/actors/ffg-character-sheet.html", "systems/starwarsffg/templates/actors/ffg-minion-sheet.html"]);
 
-  game.settings.register("starwarsffg", "systemMigrationVersion", {
-    name: "Current Version",
-    scope: "world",
-    default: null,
-    config: false,
-    type: String,
-  });
+  SettingsHelpers.initLevelSettings();
+
+  const uitheme = game.settings.get("starwarsffg", "uitheme");
+
+  switch (uitheme) {
+    case "mandar": {
+      $('link[href="systems/starwarsffg/styles/starwarsffg.css"]').prop("disabled", true);
+      $("head").append('<link href="systems/starwarsffg/styles/mandar.css" rel="stylesheet" type="text/css" media="all">');
+      break;
+    }
+    default: {
+      $('link[href="systems/starwarsffg/styles/starwarsffg.css"]').prop("disabled", false);
+    }
+  }
 
   /**
    * Set an initiative formula for the system
@@ -165,21 +172,6 @@ Hooks.once("init", async function () {
       }
     }
   }
-
-  // Register dice theme setting
-  game.settings.register("starwarsffg", "dicetheme", {
-    name: game.i18n.localize("SWFFG.SettingsDiceTheme"),
-    hint: game.i18n.localize("SWFFG.SettingsDiceThemeHint"),
-    scope: "world",
-    config: true,
-    default: "starwars",
-    type: String,
-    onChange: (rule) => window.location.reload(),
-    choices: {
-      starwars: "starwars",
-      genesys: "genesys",
-    },
-  });
 
   async function gameSkillsList() {
     game.settings.registerMenu("starwarsffg", "addskilltheme", {
@@ -292,236 +284,7 @@ Hooks.once("init", async function () {
 
   gameSkillsList();
 
-  game.settings.register("starwarsffg", "enableSoakCalc", {
-    name: game.i18n.localize("SWFFG.EnableSoakCalc"),
-    hint: game.i18n.localize("SWFFG.EnableSoakCalcHint"),
-    scope: "world",
-    config: true,
-    default: true,
-    type: Boolean,
-    onChange: (rule) => window.location.reload(),
-  });
-
-  // Register grouping talents so people can let them be ordered by purchase history
-  game.settings.register("starwarsffg", "talentSorting", {
-    name: game.i18n.localize("SWFFG.EnableSortTalentsByActivationGlobal"),
-    hint: game.i18n.localize("SWFFG.EnableSortTalentsByActivationHint"),
-    scope: "world",
-    config: true,
-    default: false,
-    type: Boolean,
-    onChange: (rule) => window.location.reload(),
-  });
-
-  // Register skill sorting by localised value setting
-  game.settings.register("starwarsffg", "skillSorting", {
-    name: game.i18n.localize("SWFFG.SettingsSkillSorting"),
-    hint: game.i18n.localize("SWFFG.SettingsSkillSortingHint"),
-    scope: "world",
-    config: true,
-    default: false,
-    type: Boolean,
-    onChange: (rule) => window.location.reload(),
-  });
-
-  // Register setting for group manager Player Character List display mode
-  game.settings.register("starwarsffg", "pcListMode", {
-    name: game.i18n.localize("SWFFG.SettingsPCListMode"),
-    hint: game.i18n.localize("SWFFG.SettingsPCListModeHint"),
-    scope: "world",
-    config: true,
-    default: "active",
-    type: String,
-    choices: {
-      active: game.i18n.localize("SWFFG.SettingsPCListModeActive"),
-      owned: game.i18n.localize("SWFFG.SettingsPCListModeOwned"),
-    },
-    onChange: (rule) => {
-      const groupmanager = canvas?.groupmanager?.window;
-      if (groupmanager) {
-        groupmanager.render();
-      }
-    },
-  });
-
-  // Register placeholder settings to store Destiny Pool values for the group manager.
-  game.settings.register("starwarsffg", "dPoolLight", {
-    name: "Destiny Pool Light",
-    scope: "world",
-    default: 0,
-    config: false,
-    type: Number,
-    onChange: (rule) => {
-      const groupmanager = canvas?.groupmanager?.window;
-      if (groupmanager) {
-        groupmanager.render();
-      }
-      let destinyLight = game.settings.get("starwarsffg", "dPoolLight");
-      document.getElementById("destinyLight").setAttribute("data-value", destinyLight);
-      document.getElementById("destinyLight").innerHTML = destinyLight + `<span>${game.i18n.localize("SWFFG.Lightside")}</span>`;
-    },
-  });
-  game.settings.register("starwarsffg", "dPoolDark", {
-    name: "Destiny Pool Dark",
-    scope: "world",
-    default: 0,
-    config: false,
-    type: Number,
-    onChange: (rule) => {
-      const groupmanager = canvas?.groupmanager?.window;
-      if (groupmanager) {
-        groupmanager.render();
-      }
-      let destinyDark = game.settings.get("starwarsffg", "dPoolDark");
-      document.getElementById("destinyDark").setAttribute("data-value", destinyDark);
-      document.getElementById("destinyDark").innerHTML = destinyDark + `<span>${game.i18n.localize("SWFFG.Darkside")}</span>`;
-    },
-  });
-
-  // Importer Control Menu
-  game.settings.registerMenu("starwarsffg", "odImporter", {
-    name: game.i18n.localize("SWFFG.SettingsOggDudeImporter"),
-    hint: game.i18n.localize("SWFFG.SettingsOggDudeImporterHint"),
-    label: game.i18n.localize("SWFFG.SettingsOggDudeImporterLabel"),
-    icon: "fas fa-file-import",
-    type: DataImporter,
-    restricted: true,
-  });
-
-  game.settings.register("starwarsffg", "odImporter", {
-    name: "Item Importer",
-    scope: "world",
-    default: {},
-    config: false,
-    default: {},
-    type: Object,
-  });
-
-  game.settings.registerMenu("starwarsffg", "swaImporter", {
-    name: game.i18n.localize("SWFFG.SettingsSWAdversariesImporter"),
-    hint: game.i18n.localize("SWFFG.SettingsSWAdversariesImporterHint"),
-    label: game.i18n.localize("SWFFG.SettingsSWAdversariesImporterLabel"),
-    icon: "fas fa-file-import",
-    type: SWAImporter,
-    restricted: true,
-  });
-
-  game.settings.register("starwarsffg", "swaImporter", {
-    name: "Adversaries Importer",
-    scope: "world",
-    default: {},
-    config: false,
-    default: {},
-    type: Object,
-  });
-
-  game.settings.register("starwarsffg", "enableDebug", {
-    name: game.i18n.localize("SWFFG.EnableDebug"),
-    hint: game.i18n.localize("SWFFG.EnableDebugHint"),
-    scope: "world",
-    config: true,
-    default: false,
-    type: Boolean,
-    onChange: (rule) => window.location.reload(),
-  });
-
-  // Set up dice with dynamic dice theme
-  const dicetheme = game.settings.get("starwarsffg", "dicetheme");
-  CONFIG.FFG.theme = dicetheme;
-
-  CONFIG.FFG.PROFICIENCY_ICON = `systems/starwarsffg/images/dice/${dicetheme}/yellow.png`;
-  CONFIG.FFG.ABILITY_ICON = `systems/starwarsffg/images/dice/${dicetheme}/green.png`;
-  CONFIG.FFG.CHALLENGE_ICON = `systems/starwarsffg/images/dice/${dicetheme}/red.png`;
-  CONFIG.FFG.DIFFICULTY_ICON = `systems/starwarsffg/images/dice/${dicetheme}/purple.png`;
-  CONFIG.FFG.BOOST_ICON = `systems/starwarsffg/images/dice/${dicetheme}/blue.png`;
-  CONFIG.FFG.SETBACK_ICON = `systems/starwarsffg/images/dice/${dicetheme}/black.png`;
-  CONFIG.FFG.REMOVESETBACK_ICON = `systems/starwarsffg/images/dice/${dicetheme}/black-minus.png`;
-  CONFIG.FFG.FORCE_ICON = `systems/starwarsffg/images/dice/${dicetheme}/whiteHex.png`;
-
-  CONFIG.FFG.ABILITY_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/green.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greens.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greens.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greenss.png'/>`, success: 2, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greena.png'/>`, success: 0, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greena.png'/>`, success: 0, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    7: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greensa.png'/>`, success: 1, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    8: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/greenaa.png'/>`, success: 0, failure: 0, advantage: 2, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-  };
-
-  CONFIG.FFG.BOOST_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blue.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blue.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blues.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/bluesa.png'/>`, success: 1, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blueaa.png'/>`, success: 0, failure: 0, advantage: 2, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/bluea.png'/>`, success: 0, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-  };
-
-  CONFIG.FFG.CHALLENGE_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/red.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redf.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redf.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redff.png'/>`, success: 0, failure: 2, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redff.png'/>`, success: 0, failure: 2, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    7: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    8: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redft.png'/>`, success: 0, failure: 1, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    9: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redft.png'/>`, success: 0, failure: 1, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    10: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redtt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 2, triumph: 0, despair: 0, light: 0, dark: 0 },
-    11: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redtt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 2, triumph: 0, despair: 0, light: 0, dark: 0 },
-    12: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/redd.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 1, light: 0, dark: 0 },
-  };
-
-  CONFIG.FFG.DIFFICULTY_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purple.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purplef.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purpleff.png'/>`, success: 0, failure: 2, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purplet.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purplet.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purplet.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    7: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purplett.png'/>`, success: 0, failure: 0, advantage: 0, threat: 2, triumph: 0, despair: 0, light: 0, dark: 0 },
-    8: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/purpleft.png'/>`, success: 0, failure: 1, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-  };
-
-  CONFIG.FFG.FORCE_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whiten.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 1 },
-    7: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitenn.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 2 },
-    8: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitel.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 1, dark: 0 },
-    9: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitel.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 1, dark: 0 },
-    10: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitell.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 2, dark: 0 },
-    11: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitell.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 2, dark: 0 },
-    12: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/whitell.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 2, dark: 0 },
-  };
-
-  CONFIG.FFG.PROFICIENCY_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellow.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellows.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellows.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowss.png'/>`, success: 2, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowss.png'/>`, success: 2, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowa.png'/>`, success: 0, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    7: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowsa.png'/>`, success: 1, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    8: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowsa.png'/>`, success: 1, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    9: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowsa.png'/>`, success: 1, failure: 0, advantage: 1, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    10: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowaa.png'/>`, success: 0, failure: 0, advantage: 2, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    11: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowaa.png'/>`, success: 0, failure: 0, advantage: 2, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    12: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/yellowr.png'/>`, success: 1, failure: 0, advantage: 0, threat: 0, triumph: 1, despair: 0, light: 0, dark: 0 },
-  };
-
-  CONFIG.FFG.SETBACK_RESULTS = {
-    1: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/black.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    2: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/black.png'/>`, success: 0, failure: 0, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    3: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blackf.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    4: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blackf.png'/>`, success: 0, failure: 1, advantage: 0, threat: 0, triumph: 0, despair: 0, light: 0, dark: 0 },
-    5: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blackt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-    6: { label: `<img src='systems/starwarsffg/images/dice/${dicetheme}/blackt.png'/>`, success: 0, failure: 0, advantage: 0, threat: 1, triumph: 0, despair: 0, light: 0, dark: 0 },
-  };
+  FFG.configureDice();
 
   // Register sheet application classes
   Actors.unregisterSheet("core", ActorSheet);
@@ -712,30 +475,7 @@ Hooks.on("renderChatMessage", (app, html, messageData) => {
 
 // Handle migration duties
 Hooks.once("ready", async () => {
-  game.settings.register("starwarsffg", "allowUsersAddRollAudio", {
-    name: game.i18n.localize("SWFFG.EnableRollAudio"),
-    hint: game.i18n.localize("SWFFG.EnableRollAudioHint"),
-    scope: "world",
-    default: false,
-    config: true,
-    type: Boolean,
-  });
-
-  const playlists = {};
-  playlists["None"] = "";
-  game.playlists.entries.forEach((playlist, index) => {
-    playlists[playlist.id] = `${index}-${playlist.data.name}`;
-  });
-
-  game.settings.register("starwarsffg", "allowUsersAddRollAudioPlaylist", {
-    name: game.i18n.localize("SWFFG.EnableRollAudioPlaylist"),
-    hint: game.i18n.localize("SWFFG.EnableRollAudioPlaylistHint"),
-    scope: "world",
-    default: "None",
-    config: true,
-    type: String,
-    choices: playlists,
-  });
+  SettingsHelpers.readyLevelSetting();
 
   const currentVersion = game.settings.get("starwarsffg", "systemMigrationVersion");
